requirements.txt
prefect==2.19.9
pyserial==3.5
adafruit-circuitpython-ds3231==2.4.20
RPi.GPIO==0.7.1
oauth2client==4.1.3
gspread==5.10.0
telepot==12.7


# python 3.9.2